added tests to OrderTest class
